// backend-cm/prisma/investigation/generated/order_type_code.ts
import { enforcement_order } from "./enforcement_order";
import { ApiProperty, ApiPropertyOptional } from "@nestjs/swagger";

export class order_type_code {
  @ApiProperty({ type: String })
  order_type_code: string;

  @ApiProperty({ type: String })
  short_description: string;

  @ApiPropertyOptional({ type: String })
  long_description?: string;

  @ApiPropertyOptional({ type: Number })
  display_order?: number;

  @ApiProperty({ type: Boolean })
  active_ind: boolean = true;

  @ApiProperty({ type: String })
  create_user_id: string;

  @ApiProperty({ type: Date })
  create_utc_timestamp: Date;

  @ApiPropertyOptional({ type: String })
  update_user_id?: string;

  @ApiPropertyOptional({ type: Date })
  update_utc_timestamp?: Date;

  @ApiProperty({ isArray: true, type: () => enforcement_order })
  enforcement_order_enforcement_order_order_type_codeToorder_type_code: enforcement_order[];
}